Improve performance by avoiding redundant shoot line calculations
https://github.com/CombatExtendedRWMod/CombatExtended/pull/129 introduced
a Harmony patch that causes drafted pawns with ranged weapons to eagerly
look for valid shoot lines to potential targets in range. Since calculating
shoot lines is very expensive, this causes massive lag when many targets
are in range but out of sight (e.g. behind a wall).

Fortunately, this patch appears to be redundant. The shoot line check added
by the patch is an eager call to CanHitTargetFrom(). If we disable the patch,
the vanilla logic (and our optimizations in Harmony_AttackTargetFinder) will
perform a set of checks including a cheap vanilla LoS calculation that immediately
rules out targets in the common case that they are completely obstructed by an obstacle.
For remaining targets, we then call AttackTargetFinder.GetRandomShootingTargetByScore(),
which then calls GetAvailableShootingTargetsByScore() -> CanShootAtFromCurrentPosition() -> CanHitTargetFrom().
Thus, we can remove our old patch and avoid expensive eager shoot line calculations.